# docs/openapi/generated/api-combined.yaml
openapi: 3.0.3
info:
  title: Weltenwind API
  version: 1.0.0
  description: |
    **Weltenwind API** - Eine moderne REST API für Multi-World Gaming Platform

    Diese API ermöglicht das Management von Benutzern, Welten, Einladungen und Themes.

    ## Authentication
    Die API verwendet JWT Bearer Token für Authentifizierung.

    ## Base URL
    - Development: `http://192.168.2.168:3000/api`
    - Production: `https://your-domain.com/api`
  contact:
    name: Weltenwind Team
    url: https://github.com/dasBoooot/weltenwind
  license:
    name: BSD 2-Clause
    url: https://opensource.org/licenses/BSD-2-Clause
servers:
  - url: http://192.168.2.168:3000/api
    description: Development Server
  - url: https://api.weltenwind.com/api
    description: Production Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token für Authentifizierung
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
...
    World:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Mittelalter-Abenteuer
        description:
          type: string
          example: Eine epische mittelalterliche Welt voller Abenteuer
        status:
          type: string
          enum:
            - upcoming
            - open
            - running
            - closed
            - archived
          example: open
        themeBundle:
          type: string
          example: medieval
        themeVariant:
          type: string
          example: castle
        playerCount:
          type: integer
          example: 42
        maxPlayers:
          type: integer
          example: 100
        createdAt:
          type: string
          format: date-time
        creator:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
...
    CreateWorldRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Meine neue Welt
        description:
          type: string
          minLength: 10
          maxLength: 1000
          example: Eine aufregende neue Welt zum Erkunden
        themeBundle:
          type: string
          example: medieval
        maxPlayers:
          type: integer
          minimum: 1
          maximum: 1000
          example: 50
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 100
        total:
          type: integer
          example: 5678
        pages:
          type: integer
          example: 57
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    CreateInviteRequest:
      type: object
      required:
        - worldId
        - email
      properties:
        worldId:
          type: integer
          example: 123
        email:
          type: string
          format: email
          example: friend@example.com
        message:
          type: string
          maxLength: 500
          example: Komm und spiele mit uns!
    Invite:
      type: object
      properties:
        id:
          type: integer
          example: 456
        token:
          type: string
          example: abc123xyz789
        worldId:
          type: integer
          example: 123
        email:
          type: string
          format: email
          example: friend@example.com
        inviterUserId:
          type: integer
          example: 789
        message:
          type: string
          example: Komm und spiele mit uns!
        status:
          type: string
          enum:
            - pending
            - accepted
            - declined
            - expired
          example: pending
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        inviteUrl:
          type: string
          example: https://weltenwind.com/invite/abc123xyz789
    InviteListItem:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        worldId:
          type: integer
        worldName:
          type: string
        message:
          type: string
        status:
          type: string
          enum:
            - pending
            - accepted
            - declined
            - expired
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        inviteUrl:
          type: string
...
    Theme:
      type: object
      properties:
        id:
          type: string
          example: medieval
        name:
          type: string
          example: Medieval
        description:
          type: string
          example: A medieval fantasy theme
        category:
          type: string
          example: fantasy
        variants:
          type: array
          items:
            type: string
          example:
            - castle
            - village
            - forest
...
    BackupOverview:
      type: object
      properties:
        config:
          $ref: '#/components/schemas/BackupConfig'
        stats:
          $ref: '#/components/schemas/BackupStats'
        health:
          $ref: '#/components/schemas/BackupHealth'
        timestamp:
          type: integer
          description: Unix timestamp der Datenerhebung
...
    QueryPerformanceOverview:
      type: object
      properties:
        totalQueries:
          type: integer
          description: Gesamtzahl der Queries
        slowQueries:
          type: integer
          description: Anzahl langsamer Queries
        avgDuration:
          type: number
          description: Durchschnittliche Query-Dauer (ms)
        medianDuration:
          type: number
          description: Median Query-Dauer (ms)
        maxDuration:
          type: number
          description: Maximale Query-Dauer (ms)
        errorRate:
          type: number
          description: Query-Fehlerquote (%)
        performance:
          $ref: '#/components/schemas/PerformanceStats'
        byOperation:
          $ref: '#/components/schemas/OperationStats'
        byTable:
          $ref: '#/components/schemas/TableStats'
        timestamp:
          type: integer
          description: Unix timestamp der Datenerhebung
        timeframeHours:
          type: integer
          description: Zeitraum der Analyse in Stunden
...
    PerformanceStats:
      type: object
      properties:
        p50:
          type: number
          description: 50. Perzentil (ms)
        p95:
          type: number
          description: 95. Perzentil (ms)
        p99:
          type: number
          description: 99. Perzentil (ms)
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_TOKEN_REQUIRED
              message: Authentication token required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: ACCESS_DENIED
              message: Insufficient permissions for this operation
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_SERVER_ERROR
              message: An internal server error occurred
paths:
...
  /worlds:
    get:
      tags:
        - Worlds
      summary: List Worlds
      description: Get a list of all available worlds
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - upcoming
              - open
              - running
              - closed
              - archived
          description: Filter by world status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
      responses:
        '200':
          description: List of worlds
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/World'
                      pagination:
                        $ref: '#/components/schemas/PaginationInfo'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Worlds
      summary: Create World
      description: Create a new world
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorldRequest'
      responses:
        '201':
          description: World created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/World'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /invites:
    post:
      tags:
        - Invites
      summary: Create Invite
      description: Create a new invitation to a world
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invite'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Invites
      summary: List My Invites
      description: Get list of invites created by current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - accepted
              - declined
              - expired
          description: Filter by invite status
        - name: worldId
          in: query
          schema:
            type: integer
          description: Filter by world ID
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/InviteListItem'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /themes:
    get:
      tags:
        - Themes
      summary: List Available Themes
      description: Get list of all available themes
      responses:
        '200':
          description: List of themes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Theme'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/backup:
    get:
      summary: Backup-System-Übersicht
      description: Liefert umfassende Informationen über das Backup-System, Konfiguration, Statistiken und Health-Status
      tags:
        - Backup Management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Erfolgreich - Backup-System-Übersicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupOverview'
        '401':
          description: Unauthorized - Nicht authentifiziert
        '403':
          description: Forbidden - Keine Admin-Berechtigung
        '500':
          description: Interner Serverfehler
...
  /api/query-performance:
    get:
      summary: Query-Performance-Übersicht
      description: Umfassende Statistiken über Datenbank-Query-Performance mit Performance-Metriken
      tags:
        - Query Performance
      security:
        - bearerAuth: []
      parameters:
        - name: timeframe
          in: query
          description: Zeitraum für Statistiken in Stunden (optional)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
      responses:
        '200':
          description: Erfolgreich - Query-Performance-Daten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryPerformanceOverview'
        '401':
          description: Unauthorized - Nicht authentifiziert
        '403':
          description: Forbidden - Keine Admin-Berechtigung
        '500':
          description: Interner Serverfehler
...